test: isolate CLI flush integration
