+++ test/integration/daemon-replace-exit-flush.test.ts
@@ -4,11 +4,9 @@ import fs from 'node:fs';
 import os from 'node:os';
 import path from 'node:path';
 import { skipWhenLoopbackUnavailable } from '../../src/__tests__/test-utils/loopback.ts';
-import { runCmd } from '../../src/utils/exec.ts';
 import { stopProcessForTakeover } from '../../src/daemon/daemon-process.ts';
 import { isProcessAlive } from '../../src/utils/host-process.ts';
 import { runCliJson } from './test-helpers.ts';
-import { PAYLOAD_MARKER } from './support/exit-payload.ts';
 
 // #1596: a CLI command that finds its recorded daemon unreachable replaces it
 // (`Replacing daemon (pid N, vX) in <state-dir>: unreachable`) and retries
@@ -17,9 +15,6 @@ import { PAYLOAD_MARKER } from './support/exit-payload.ts';
 // has no sessions yet, which is expected). This file locks down that a
 // replace-mid-command always ends in a normal, fully-delivered structured
 // error rather than a truncated or hung process.
-
-const SUPPORT_DIR = path.join(import.meta.dirname, 'support');
-const FIXTURE_TIMEOUT_MS = 10_000;
 
 type DaemonInfo = {
   pid: number;
@@ -85,39 +80,6 @@ test('daemon replace mid-command returns a structured, parseable error and exits
     fs.rmSync(stateDir, { recursive: true, force: true });
   }
 });
-
-// Isolates the delivery guarantee from the end-to-end test above. The fixture
-// writes a large payload through a real piped subprocess without asserting
-// that the unsafe alternative must truncate under a particular OS, Node
-// version, pipe configuration, or scheduler interleaving.
-test('exitAfterFlush (the #1596 fix) delivers the full write before the process exits', async () => {
-  const { exitCode, stderr } = await runFixture('exit-after-flush.ts');
-  assert.equal(exitCode, 1);
-  assert.ok(
-    stderr.includes(PAYLOAD_MARKER),
-    'expected the full payload, including its trailing marker',
-  );
-});
-
-test('the CLI success entrypoint flushes a complete JSON envelope before exit 0', async () => {
-  const { exitCode, stdout } = await runFixture('exit-after-flush.ts', ['--success']);
-  assert.equal(exitCode, 0);
-  const envelope = JSON.parse(stdout) as { success?: boolean; data?: { payload?: string } };
-  assert.equal(envelope.success, true);
-  assert.equal(envelope.data?.payload?.endsWith(PAYLOAD_MARKER), true);
-});
-
-async function runFixture(
-  name: string,
-  args: string[] = [],
-): Promise<{ exitCode: number; stdout: string; stderr: string }> {
-  const result = await runCmd(
-    process.execPath,
-    ['--experimental-strip-types', path.join(SUPPORT_DIR, name), ...args],
-    { allowFailure: true, timeoutMs: FIXTURE_TIMEOUT_MS },
-  );
-  return { exitCode: result.exitCode, stdout: result.stdout, stderr: result.stderr };
-}
 
 async function waitForProcessDeath(pid: number): Promise<void> {
   const deadline = Date.now() + 5_000;
